+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 0.2.0.0 (2021-08-31)
+
+- Add a type family `IsSat :: Constraint -> Bool`
+  that computes whether a type-family is satisfied in
+  the current context.
+
+- Rename `IfCt` to `IfSat`.
+
 # Version 0.1.0.0 (2021-08-30)
 
 Initial release.